Refactor code to utilize react-query for fetching and adding tasks
