+++ client/src/components/TaskModal.tsx
@@ -1,9 +1,8 @@
-import type { Task } from '../common/Task';
 import type { TaskInput } from '../common/TaskInput';
 
 import { useState } from 'react';
-import axios from 'axios';
-import { useMutation } from '@tanstack/react-query';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { addTask } from '../api/taskapi';
 
 interface TaskModalProps {
   closeTaskModal: () => void;
@@ -11,24 +10,11 @@ interface TaskModalProps {
 }
 
 export function TaskModal(props: TaskModalProps) {
+  const queryClient = useQueryClient();
+
   const [taskTitle, setTaskTitle] = useState('');
   const [taskDetails, setTaskDetails] = useState('');
   const [taskPoints, setTaskPoints] = useState<number>(0);
-
-  const addTask = async(input: TaskInput) => {
-    try {
-      const config = {
-        method: 'post',
-        baseURL: import.meta.env.VITE_REACT_APP_SERVER,
-        url: '/task',
-        data: input,
-      };
-      const result = await axios<Task>(config);
-      return result;
-    } catch (err) {
-      console.error(err);
-    }
-  };
 
   function addTaskHandler() {
     const newTask: TaskInput = {
@@ -36,11 +22,19 @@ export function TaskModal(props: TaskModalProps) {
       content: taskDetails,
       point_value: taskPoints,
     };
+    addTaskMutate(newTask);
     setTaskTitle('');
     setTaskDetails('');
     setTaskPoints(0);
     props.closeTaskModal();
   }
+
+  const {isError, error, mutate: addTaskMutate} = useMutation({
+    mutationFn: addTask,
+    onSuccess: () => {
+      queryClient.invalidateQueries({queryKey: ['get-tasks']});
+    },
+  });
   
   return (
     <>
@@ -64,6 +58,10 @@ export function TaskModal(props: TaskModalProps) {
             </label>
             <div id="task-modal-btns">
               <button onClick={addTaskHandler}>Add New Task</button>
+
+              {isError &&
+                <p>{error?.message}</p>
+              }
             </div>
           </div>
         </div>
